requirements.txt
treelib
python-dotenv
PyQt6
rich